build, validation, test: treat nVersion as signed and add tests

+++ src/validation.c
@@ -50,15 +50,15 @@ dogecoin_bool dogecoin_block_header_scrypt_hash(cstring* s, uint256_t* hash) {
     return true;
 }
 
-uint32_t get_chainid(uint32_t version) {
-    return version >> 16;
+uint32_t get_chainid(int32_t version) {
+    return ((uint32_t)version) >> 16;
 }
 
-dogecoin_bool is_auxpow(uint32_t version) {
-    return (version & (1 << 8)) == 256;
+dogecoin_bool is_auxpow(int32_t version) {
+    return (((uint32_t)version) & (1 << 8)) == 256;
 }
 
-dogecoin_bool is_legacy(uint32_t version) {
+dogecoin_bool is_legacy(int32_t version) {
     return version == 1
         // Dogecoin: We have a random v2 block with no AuxPoW, treat as legacy
         || (version == 2 && get_chainid(version) == 0);
